SQL-2161: create cli with clap and confy (#733)
* SQL-2161: create cli with clap and confy

This is the basic skeleton of hte cli with all arguments. The config file is added for now for reference, and will be expanded upon and more thoroughly tested closer to when we call the builder library from the tool.

I did add a test to cover the merge function which will allow cli args to override file args

* fix dep name

* feedback

---------

Co-authored-by: Nathan Leniz <nathanleniz@mongodb.com>