SQL-2001: gracefully handle bindata conversion to string for all versions of MongoDB (#814)
* SQL-2001: gracefully handle bindata conversion to string for all versions of MongoDB

This modifies the codegen output of $convert when the `input` type is `"binData"` and the `to` type is `"string"`. This is required because in versions of MongoDB < 8.0, this conversion isn't supported so we benefit from the onError handling, which we specify to be `null`.

However, beginning with MongoDB 8.0, ths type of conversion _is_ supported. However, when the `input` type is `"binData"` and the `to` type is `"string"`, `format` is *required* and it fails to parse, which does not trigger the `onError` handling and instead returns an error to the user.

Until we can better support additional `CastExpr` options (ticket to be created), this may be the best we can do, and will likely be what we continue to do when users do not specify the new options once they are added.

* update codegen tests

* remove comment template

* feedback

* update test